+++ src/app/components/base-filter-page/index.ts
@@ -1,0 +1,1 @@
+export * from './base-filter-page.component';
